refactor: improve timekeeping
